Initial implementation of Chirp chat
Made the foundation for Chirp handling.

Made regex for handling the CSV file and using it with the MatchCollection to separate the given line into 3 parts. Then putting this into a list we can use to declare variables we then can use to make Chirps.

We also made a StreamWriter which handles user input which takes the message, author and timestamp and appends it into the CSV file.

Lastly we made a Chirp class which contains the 3 different variables. Here we have a constructor which saves this information in each individual Chirp.

Then it has a separat print function which calls the formatTimeStamp function which formats the time stamp to a DateTime, which then is returned to the print function, that returns a string, in order to ensure that the text printed in the terminal matches the layout of the assignment.