We removed everything, and made a razor app
This branch is reserved to working on- and setting up the razor app. When completed, the branch will be merged with main.

We did this using the following command:
dotnet dotnet new chirp-razor -o Chirp.Razor

The reason behind this change is described in the README_PROJECT.md file under session_05 in lecture notes.

Co-Authored-By: KongCristi <151643927+KongCristi@users.noreply.github.com>
Co-Authored-By: Marcus Egede Frandsen <megf@itu.dk>
Co-Authored-By: benjaminormstrup <151547859+benjaminormstrup@users.noreply.github.com>