Seeding the database
Implemented the DbInitializer.cs to seed the database.

We have updated the pages folder files.

We have updated the repositories folder files so the correct template is used. Methods will be added later this week.

We have errors due to lack of methods at the moment, but as stated, this will be looked at later this week.

Co-Authored-By: KongCristi <151643927+KongCristi@users.noreply.github.com>
Co-Authored-By: McLarsenITU <150607179+McLarsenITU@users.noreply.github.com>
Co-Authored-By: benjaminormstrup <151547859+benjaminormstrup@users.noreply.github.com>
Co-Authored-By: victorroep <15931564+victorroep@users.noreply.github.com>