Pair Programming Switch
Made a new project with "dotnet new webapp -au Individual" to see what was made. Then modified the cshtml file to have the navigation bar as described in the project task. Made a css file to enable the future buttons to align horisontally with the correct amount of space between.

Co-Authored-By: KongCristi <151643927+KongCristi@users.noreply.github.com>
Co-Authored-By: McLarsenITU <150607179+McLarsenITU@users.noreply.github.com>