Cheeps posting modified
Now the cheepbox has the ability to auto resize. The char counter stays in the bottom right of the cheepbox and the "Share" button stays alligned with the top of the cheepbox.

Once a cheep was posted, we had a problem with long words exceeding the border of the post. This has now been fixed to break words so it no longer can exceed this box.

This applies to both public and private timelines

Co-Authored-By: KongCristi <151643927+KongCristi@users.noreply.github.com>
Co-Authored-By: McLarsenITU <150607179+McLarsenITU@users.noreply.github.com>
Co-Authored-By: benjaminormstrup <151547859+benjaminormstrup@users.noreply.github.com>